chore: improve batch actions
